Refactor mouse scrolling code
